convert coolant control and actually configure the io pins

+++ coolant_control.h
@@ -21,7 +21,7 @@
 #ifndef coolant_control_h
 #define coolant_control_h 
 
-#include <avr/io.h>
+#include <stdint.h>
 
 #define COOLANT_MIST_ENABLE 2
 #define COOLANT_FLOOD_ENABLE 1
@@ -29,6 +29,6 @@
 
 void coolant_init();
 void coolant_stop();
-void coolant_run(uint8_t mode);
+void coolant_run(uint32_t mode);
 
 #endif
